app/requirements.txt
pandas
scikit-learn
joblib
psycopg2-binary
matplotlib
numpy
jinja2
